December 2011 version as-is, no change whatsoever
This is my FIRST java/programming project EVER !!
I followed a class as a non-student back at the university,
"introduction to programming (in java)". I knew *nothing* about
computers, struggling to open word... And there it is, after a month
of playing around (in Java 6), I came up with that. I was soooo proud.

Anyway, here is it, a relic of my debug in programming :)